Clear compiler warnings inherited from Arduino code use.

+++ hardware/arduino/cores/arduino/Stream.cpp
@@ -194,7 +194,7 @@ float Stream::parseFloat(char skipChar){
     else if(c >= '0' && c <= '9')  {      // is c a digit?
       value = value * 10 + c - '0';
       if(isFraction)
-         fraction *= 0.1;
+         fraction *= 0.1f;
     }
     read();  // consume the character we got with peek
     c = timedPeek();
@@ -206,7 +206,7 @@ float Stream::parseFloat(char skipChar){
   if(isFraction)
     return value * fraction;
   else
-    return value;
+    return static_cast<float>(value);
 }
 
 // read characters from stream into buffer
